@@ -9,24 +9,24 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "canvas-confetti": "^1.9.2",
+    "canvas-confetti": "^1.9.3",
     "clipboard": "^2.0.11",
     "mitt": "^3.0.1",
     "nes.css": "^2.3.0",
-    "pinia": "^2.1.7",
-    "socket.io-client": "^4.7.5",
+    "pinia": "^2.2.4",
+    "socket.io-client": "^4.8.0",
     "velocity-animate": "^2.0.6",
-    "vue": "^3.4.15",
-    "vue-router": "^4.3.0"
+    "vue": "^3.5.12",
+    "vue-router": "^4.4.5"
   },
   "devDependencies": {
-    "@iconify-json/pixelarticons": "^1.1.10",
+    "@iconify-json/pixelarticons": "^1.2.1",
     "@iconify/tailwind": "^0.1.4",
-    "@vitejs/plugin-vue": "^5.0.3",
-    "autoprefixer": "^10.4.18",
-    "postcss": "^8.4.35",
-    "tailwindcss": "^3.4.1",
-    "vite": "^5.0.11"
+    "@vitejs/plugin-vue": "^5.1.4",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.14",
+    "vite": "^5.4.10"
   },
   "packageManager": "pnpm@9.12.1+sha512.e5a7e52a4183a02d5931057f7a0dbff9d5e9ce3161e33fa68ae392125b79282a8a8a470a51dfc8a0ed86221442eb2fb57019b0990ed24fab519bf0e1bc5ccfc4"
 }
