@@ -1,4 +1,4 @@
-const { addDynamicIconSelectors } = require('@iconify/tailwind')
+import { addDynamicIconSelectors } from '@iconify/tailwind'
 
 /** @type {import('tailwindcss').Config} */
 export default {
